Added UIImage Caching; Refactored Download Image Network Call
Added a publicly accessible cache = NSCache<NSString, UIImage>() in the NetworkManager to use for reusing already download images for the GUAvatarImageView. This will allow for faster scrolling and prevent stoppage related to downloading images that have already been used once over the network again. Also refactored the downloadImage() to be in the NetworkManager, previously had it in the GUAvatarImageView due to less overall code needing to be implemeneted, however decided that I would prioritize organization and have all network calls be done through the network manager.
