+++ GithubUsers/GithubUsers/Custom Views/Cells/FollowerCell.swift
@@ -26,7 +26,7 @@ class FollowerCell: UICollectionViewCell {
     
     func set(follower: Follower) {
         usernameLabel.text = follower.login
-        avatarImageView.downloadImage(from: follower.avatarUrl)
+        avatarImageView.setImage(from: follower.avatarUrl)
     }
     
     private func configure() {
