+++ GithubUsers/GithubUsers/Models/NetworkManager.swift
@@ -5,12 +5,13 @@
 //  Created by Marco Capraro on 2/15/24.
 //
 
-import Foundation
+import UIKit
 
 // Basic native way to make network calls (foundational knowledge)
 class NetworkManager {
     static let shared = NetworkManager()
-    let baseURL = "https://api.github.com/users/"
+    private let baseURL = "https://api.github.com/users/"
+    let cache = NSCache<NSString, UIImage>()
     
     private init() {}
     
@@ -65,6 +66,34 @@ class NetworkManager {
         }
         
         // This starts the network call created above
+        task.resume()
+    }
+    
+    func downloadImage(from urlString: String, completed: @escaping (UIImage) -> Void) {
+        let cacheKey = NSString(string: urlString)
+        
+        if let image = cache.object(forKey: cacheKey) {
+            completed(image)
+            return
+        }
+        
+        guard let url = URL(string: urlString) else { return }
+        
+        let task = URLSession.shared.dataTask(with: url) { [weak self] data, response, error in
+            guard let self = self else { return }
+            
+            if error != nil { return }
+            guard let response = response as? HTTPURLResponse, response.statusCode == 200 else { return }
+            guard let data = data else { return }
+            
+            guard let image = UIImage(data: data) else { return }
+            self.cache.setObject(image, forKey: cacheKey)
+            
+            DispatchQueue.main.async {
+                completed(image)
+            }
+        }
+        
         task.resume()
     }
 }
